+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 - Add support for PAYG Instalments, which are specified in aggregate
   as a refundable tax offset via the `paygInstalments` field.
 
+- Add the `Deductions` type, which expresses the various deduction
+  types, as well as the (aggregate) amount of deductions related to
+  foreign income.
+
 ## Older versions
 
 See Git commit history